fix font stuff, docs

+++ src/core/shared/render.cpp
@@ -60,7 +60,7 @@ sdk::vec2_t c_render::get_screen_size() {
 	return sdk::vec2_t(w, h);
 }
 
-sdk::h_font c_render::get_font(int id) {
+sdk::h_font c_render::get_font(unsigned long id) {
 	auto default_font = wh->portal2->scheme->get_font("DefaultFixedOutline");
 
 	return default_font + id;
